[CORE-1041] Expand RealEnv Mock and Convert Auth Integration Tests to Unit Tests (#8163)
