+++ src/internal/metrics/metrics.go
@@ -156,7 +156,7 @@ func (r *Reporter) reportClusterMetrics() {
 	}
 }
 
-func externalMetrics(kubeClient *kube.Clientset, metrics *Metrics) error {
+func externalMetrics(kubeClient kube.Interface, metrics *Metrics) error {
 	nodeList, err := kubeClient.CoreV1().Nodes().List(context.Background(), metav1.ListOptions{})
 	if err != nil {
 		return errors.Wrapf(err, "externalMetrics: unable to retrieve node list from k8s")
