+++ src/server/pfs/s3/master_test.go
@@ -1,9 +1,11 @@
-package s3
+package s3_test
 
 import (
 	"fmt"
 	"io"
-	
+
+	"github.com/pachyderm/pachyderm/v2/src/server/pfs/s3"
+
 	"os"
 	"strings"
 	"testing"
@@ -532,7 +534,7 @@ func TestMasterDriver(t *testing.T) {
 	}
 	t.Parallel()
 	env := testpachd.NewRealEnv(t, dockertestenv.NewTestDBConfig(t))
-	testRunner(t, env.PachClient, "master", NewMasterDriver(), func(t *testing.T, pachClient *client.APIClient, minioClient *minio.Client) {
+	testRunner(t, env.PachClient, "master", s3.NewMasterDriver(), func(t *testing.T, pachClient *client.APIClient, minioClient *minio.Client) {
 		t.Run("ListBuckets", func(t *testing.T) {
 			masterListBuckets(t, pachClient, minioClient)
 		})
