+++ src/server/pfs/s3/util_test.go
@@ -1,11 +1,13 @@
-package s3
+package s3_test
 
 import (
 	"context"
 	"crypto/md5"
 	"fmt"
 	"io"
-	
+
+	"github.com/pachyderm/pachyderm/v2/src/server/pfs/s3"
+
 	"net"
 	"os"
 	"path/filepath"
@@ -129,11 +131,11 @@ func fileHash(t *testing.T, name string) (int64, []byte) {
 	return fi.Size(), hashSum
 }
 
-func testRunner(t *testing.T, pachClient *client.APIClient, group string, driver Driver, runner func(t *testing.T, pachClient *client.APIClient, minioClient *minio.Client)) {
-	router := Router(driver, func(_ctx context.Context) *client.APIClient {
+func testRunner(t *testing.T, pachClient *client.APIClient, group string, driver s3.Driver, runner func(t *testing.T, pachClient *client.APIClient, minioClient *minio.Client)) {
+	router := s3.Router(driver, func(_ctx context.Context) *client.APIClient {
 		return pachClient.WithCtx(context.Background())
 	})
-	server := Server(0, router)
+	server := s3.Server(0, router)
 	listener, err := net.Listen("tcp", ":0")
 	require.NoError(t, err)
