+++ src/server/pps/server/infra_driver.go
@@ -31,49 +31,49 @@ type InfraDriver interface {
 	WatchPipelinePods(ctx context.Context) (<-chan watch.Event, func(), error)
 }
 
-type mockInfraOp int32
+type MockInfraOp int32
 
 const (
-	mockInfraOp_CREATE = 0
-	mockInfraOp_DELETE = 1
-	mockInfraOp_READ   = 2
+	MockInfraOp_CREATE = 0
+	MockInfraOp_DELETE = 1
+	MockInfraOp_READ   = 2
 )
 
-type mockInfraDriver struct {
-	rcs          map[string]v1.ReplicationController // indexed by pipeline name
-	calls        map[string]map[mockInfraOp]int      // indexed by pipeline name
-	scaleHistory map[string][]int32                  // indexed by pipeline name
+type MockInfraDriver struct {
+	Rcs          map[string]v1.ReplicationController // indexed by pipeline name
+	Calls        map[string]map[MockInfraOp]int      // indexed by pipeline name
+	ScaleHistory map[string][]int32                  // indexed by pipeline name
 }
 
-func newMockInfraDriver() *mockInfraDriver {
-	d := &mockInfraDriver{
-		rcs:          make(map[string]v1.ReplicationController),
-		calls:        make(map[string]map[mockInfraOp]int),
-		scaleHistory: make(map[string][]int32),
+func NewMockInfraDriver() *MockInfraDriver {
+	d := &MockInfraDriver{
+		Rcs:          make(map[string]v1.ReplicationController),
+		Calls:        make(map[string]map[MockInfraOp]int),
+		ScaleHistory: make(map[string][]int32),
 	}
 	return d
 }
 
-func (d *mockInfraDriver) CreatePipelineResources(ctx context.Context, pi *pps.PipelineInfo) error {
-	d.rcs[pi.Pipeline.Name] = *d.makeRC(pi)
-	d.incCall(pi.Pipeline.Name, mockInfraOp_CREATE)
-	if _, ok := d.scaleHistory[pi.Pipeline.Name]; !ok {
-		d.scaleHistory[pi.Pipeline.Name] = make([]int32, 0)
+func (d *MockInfraDriver) CreatePipelineResources(ctx context.Context, pi *pps.PipelineInfo) error {
+	d.Rcs[pi.Pipeline.Name] = *d.MakeRC(pi)
+	d.IncCall(pi.Pipeline.Name, MockInfraOp_CREATE)
+	if _, ok := d.ScaleHistory[pi.Pipeline.Name]; !ok {
+		d.ScaleHistory[pi.Pipeline.Name] = make([]int32, 0)
 	}
 	return nil
 }
 
-func (d *mockInfraDriver) DeletePipelineResources(ctx context.Context, pipeline string) error {
-	if _, ok := d.rcs[pipeline]; ok {
-		delete(d.rcs, pipeline)
-		d.incCall(pipeline, mockInfraOp_DELETE)
+func (d *MockInfraDriver) DeletePipelineResources(ctx context.Context, pipeline string) error {
+	if _, ok := d.Rcs[pipeline]; ok {
+		delete(d.Rcs, pipeline)
+		d.IncCall(pipeline, MockInfraOp_DELETE)
 	}
 	return nil
 }
 
-func (d *mockInfraDriver) ReadReplicationController(ctx context.Context, pi *pps.PipelineInfo) (*v1.ReplicationControllerList, error) {
-	d.incCall(pi.Pipeline.Name, mockInfraOp_READ)
-	if rc, ok := d.rcs[pi.Pipeline.Name]; !ok {
+func (d *MockInfraDriver) ReadReplicationController(ctx context.Context, pi *pps.PipelineInfo) (*v1.ReplicationControllerList, error) {
+	d.IncCall(pi.Pipeline.Name, MockInfraOp_READ)
+	if rc, ok := d.Rcs[pi.Pipeline.Name]; !ok {
 		return &v1.ReplicationControllerList{
 			Items: []v1.ReplicationController{},
 		}, errors.New("rc for pipeline not found")
@@ -84,19 +84,19 @@ func (d *mockInfraDriver) ReadReplicationController(ctx context.Context, pi *pps
 	}
 }
 
-func (d *mockInfraDriver) UpdateReplicationController(ctx context.Context, old *v1.ReplicationController, update func(rc *v1.ReplicationController) bool) error {
+func (d *MockInfraDriver) UpdateReplicationController(ctx context.Context, old *v1.ReplicationController, update func(rc *v1.ReplicationController) bool) error {
 	rc := old.DeepCopy()
 	if update(rc) {
 		name := rc.ObjectMeta.Labels[pipelineNameLabel]
-		d.scaleHistory[name] = append(d.scaleHistory[name], *rc.Spec.Replicas)
-		d.writeRC(rc)
+		d.ScaleHistory[name] = append(d.ScaleHistory[name], *rc.Spec.Replicas)
+		d.WriteRC(rc)
 	}
 	return nil
 }
 
-func (d *mockInfraDriver) ListReplicationControllers(ctx context.Context) (*v1.ReplicationControllerList, error) {
+func (d *MockInfraDriver) ListReplicationControllers(ctx context.Context) (*v1.ReplicationControllerList, error) {
 	items := make([]v1.ReplicationController, 0)
-	for _, rc := range d.rcs {
+	for _, rc := range d.Rcs {
 		items = append(items, rc)
 	}
 	return &v1.ReplicationControllerList{
@@ -105,7 +105,7 @@ func (d *mockInfraDriver) ListReplicationControllers(ctx context.Context) (*v1.R
 }
 
 // TODO(acohen4): complete
-func (d *mockInfraDriver) WatchPipelinePods(ctx context.Context) (<-chan watch.Event, func(), error) {
+func (d *MockInfraDriver) WatchPipelinePods(ctx context.Context) (<-chan watch.Event, func(), error) {
 	ch := make(chan watch.Event)
 	return ch, func() {}, nil
 }
@@ -114,11 +114,11 @@ func (d *mockInfraDriver) WatchPipelinePods(ctx context.Context) (<-chan watch.E
 // -------- Mock Helpers -------- //
 ////////////////////////////////////
 
-func (d *mockInfraDriver) resetRCs() {
-	d.rcs = make(map[string]v1.ReplicationController)
+func (d *MockInfraDriver) ResetRCs() {
+	d.Rcs = make(map[string]v1.ReplicationController)
 }
 
-func (d *mockInfraDriver) makeRC(pi *pps.PipelineInfo) *v1.ReplicationController {
+func (d *MockInfraDriver) MakeRC(pi *pps.PipelineInfo) *v1.ReplicationController {
 	return &v1.ReplicationController{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: ppsutil.PipelineRcName(pi.Pipeline.Name, pi.Version),
@@ -133,15 +133,15 @@ func (d *mockInfraDriver) makeRC(pi *pps.PipelineInfo) *v1.ReplicationController
 	}
 }
 
-func (d *mockInfraDriver) writeRC(rc *v1.ReplicationController) {
+func (d *MockInfraDriver) WriteRC(rc *v1.ReplicationController) {
 	name := rc.ObjectMeta.Labels[pipelineNameLabel]
-	d.rcs[name] = *rc
+	d.Rcs[name] = *rc
 }
 
-func (d *mockInfraDriver) incCall(name string, op mockInfraOp) {
-	if ops, ok := d.calls[name]; ok {
+func (d *MockInfraDriver) IncCall(name string, op MockInfraOp) {
+	if ops, ok := d.Calls[name]; ok {
 		ops[op] += 1
 	} else {
-		d.calls[name] = map[mockInfraOp]int{op: 1}
+		d.Calls[name] = map[MockInfraOp]int{op: 1}
 	}
 }
