+++ src/server/pps/server/kube_driver.go
@@ -20,7 +20,7 @@ import (
 )
 
 type kubeDriver struct {
-	kubeClient *kubernetes.Clientset
+	kubeClient kubernetes.Interface
 	namespace  string
 	// the limiter intends to guard the k8s API server from being overwhelmed by many concurrent requests
 	// that could arise from many concurrent pipelineController goros.
@@ -30,7 +30,7 @@ type kubeDriver struct {
 	logger     *logrus.Logger
 }
 
-func newKubeDriver(kubeClient *kubernetes.Clientset, config serviceenv.Configuration, logger *logrus.Logger) InfraDriver {
+func newKubeDriver(kubeClient kubernetes.Interface, config serviceenv.Configuration, logger *logrus.Logger) InfraDriver {
 	return &kubeDriver{
 		kubeClient: kubeClient,
 		namespace:  config.Namespace,
