+++ src/server/worker/datum/iterator_test.go
@@ -1,4 +1,4 @@
-package datum
+package datum_test
 
 import (
 	"context"
@@ -7,6 +7,8 @@ import (
 	"sort"
 	"strings"
 	"testing"
+
+	"github.com/pachyderm/pachyderm/v2/src/server/worker/datum"
 
 	"github.com/gogo/protobuf/types"
 	"github.com/pachyderm/pachyderm/v2/src/client"
@@ -38,7 +40,7 @@ func TestIterators(t *testing.T) {
 	in0 := client.NewPFSInput(dataRepo, "!(**)")
 	in0.Pfs.Commit = commit.ID
 	t.Run("ZeroDatums", func(t *testing.T) {
-		pfs0, err := NewIterator(c, taskDoer, in0)
+		pfs0, err := datum.NewIterator(c, taskDoer, in0)
 		require.NoError(t, err)
 		validateDI(t, pfs0)
 	})
@@ -48,9 +50,9 @@ func TestIterators(t *testing.T) {
 	in2 := client.NewPFSInput(dataRepo, "/foo*2")
 	in2.Pfs.Commit = commit.ID
 	t.Run("Basic", func(t *testing.T) {
-		pfs1, err := NewIterator(c, taskDoer, in1)
+		pfs1, err := datum.NewIterator(c, taskDoer, in1)
 		require.NoError(t, err)
-		pfs2, err := NewIterator(c, taskDoer, in2)
+		pfs2, err := datum.NewIterator(c, taskDoer, in2)
 		require.NoError(t, err)
 		validateDI(t, pfs1, "/foo11", "/foo21", "/foo31", "/foo41")
 		validateDI(t, pfs2, "/foo12", "/foo2", "/foo22", "/foo32", "/foo42")
@@ -58,14 +60,14 @@ func TestIterators(t *testing.T) {
 	// Union input.
 	in3 := client.NewUnionInput(in1, in2)
 	t.Run("Union", func(t *testing.T) {
-		union1, err := NewIterator(c, taskDoer, in3)
+		union1, err := datum.NewIterator(c, taskDoer, in3)
 		require.NoError(t, err)
 		validateDI(t, union1, "/foo11", "/foo12", "/foo2", "/foo21", "/foo22", "/foo31", "/foo32", "/foo41", "/foo42")
 	})
 	// Cross input.
 	in4 := client.NewCrossInput(in1, in2)
 	t.Run("Cross", func(t *testing.T) {
-		cross1, err := NewIterator(c, taskDoer, in4)
+		cross1, err := datum.NewIterator(c, taskDoer, in4)
 		require.NoError(t, err)
 		validateDI(t, cross1,
 			"/foo11/foo12", "/foo11/foo2", "/foo11/foo22", "/foo11/foo32", "/foo11/foo42",
@@ -77,14 +79,14 @@ func TestIterators(t *testing.T) {
 	// Empty cross.
 	in6 := client.NewCrossInput(in3, in0, in2, in4)
 	t.Run("EmptyCross", func(t *testing.T) {
-		cross3, err := NewIterator(c, taskDoer, in6)
+		cross3, err := datum.NewIterator(c, taskDoer, in6)
 		require.NoError(t, err)
 		validateDI(t, cross3)
 	})
 	// Nested empty cross.
 	in7 := client.NewCrossInput(in6, in1)
 	t.Run("NestedEmptyCross", func(t *testing.T) {
-		cross4, err := NewIterator(c, taskDoer, in7)
+		cross4, err := datum.NewIterator(c, taskDoer, in7)
 		require.NoError(t, err)
 		validateDI(t, cross4)
 	})
@@ -95,7 +97,7 @@ func TestIterators(t *testing.T) {
 	in9.Pfs.Commit = commit.ID
 	in10 := client.NewJoinInput(in8, in9)
 	t.Run("Join", func(t *testing.T) {
-		join1, err := NewIterator(c, taskDoer, in10)
+		join1, err := datum.NewIterator(c, taskDoer, in10)
 		require.NoError(t, err)
 		validateDI(t, join1,
 			"/foo11/foo11",
@@ -122,7 +124,7 @@ func TestIterators(t *testing.T) {
 	in12.Pfs.Commit = commit.ID
 	in13 := client.NewJoinInput(in11, in12)
 	t.Run("OuterJoin", func(t *testing.T) {
-		join1, err := NewIterator(c, taskDoer, in13)
+		join1, err := datum.NewIterator(c, taskDoer, in13)
 		require.NoError(t, err)
 		validateDI(t, join1,
 			"/foo10",
@@ -141,7 +143,7 @@ func TestIterators(t *testing.T) {
 	in14.Pfs.Commit = commit.ID
 	in15 := client.NewGroupInput(in14)
 	t.Run("GroupSingle", func(t *testing.T) {
-		group1, err := NewIterator(c, taskDoer, in15)
+		group1, err := datum.NewIterator(c, taskDoer, in15)
 		require.NoError(t, err)
 		validateDI(t, group1,
 			"/foo10/foo11/foo12/foo13/foo14/foo15/foo16/foo17/foo18/foo19",
@@ -156,7 +158,7 @@ func TestIterators(t *testing.T) {
 	in17.Pfs.Commit = commit.ID
 	in18 := client.NewGroupInput(in16, in17)
 	t.Run("GroupDoubles", func(t *testing.T) {
-		group2, err := NewIterator(c, taskDoer, in18)
+		group2, err := datum.NewIterator(c, taskDoer, in18)
 		require.NoError(t, err)
 		validateDI(t, group2,
 			"/foo10/foo20/foo30/foo40",
@@ -179,7 +181,7 @@ func TestIterators(t *testing.T) {
 	in21 := client.NewJoinInput(in19, in20)
 	in22 := client.NewGroupInput(in21)
 	t.Run("GroupJoin", func(t *testing.T) {
-		groupJoin1, err := NewIterator(c, taskDoer, in22)
+		groupJoin1, err := datum.NewIterator(c, taskDoer, in22)
 		require.NoError(t, err)
 		validateDI(t, groupJoin1,
 			"/foo11/foo11/foo12/foo21/foo13/foo31/foo14/foo41",
@@ -197,7 +199,7 @@ func TestIterators(t *testing.T) {
 	in26 := client.NewUnionInput(in23, in25)
 
 	t.Run("UnionGroup", func(t *testing.T) {
-		unionGroup1, err := NewIterator(c, taskDoer, in26)
+		unionGroup1, err := datum.NewIterator(c, taskDoer, in26)
 		require.NoError(t, err)
 		validateDI(t, unionGroup1,
 			"/foo10/foo20/foo30/foo40",
@@ -256,11 +258,11 @@ func TestIterators(t *testing.T) {
 	in27 := client.NewS3PFSInput("", dataRepo, "")
 	in27.Pfs.Commit = commit.ID
 	t.Run("PlainS3", func(t *testing.T) {
-		di, err := NewIterator(c, taskDoer, in27)
+		di, err := datum.NewIterator(c, taskDoer, in27)
 		require.NoError(t, err)
 		validateDI(t, di, "/")
 		// Check that every datum has an S3 input
-		require.NoError(t, di.Iterate(func(meta *Meta) error {
+		require.NoError(t, di.Iterate(func(meta *datum.Meta) error {
 			require.True(t, meta.Inputs[0].S3)
 			return nil
 		}))
@@ -269,7 +271,7 @@ func TestIterators(t *testing.T) {
 	// in28 is a cross that contains an S3 input and two non-s3 inputs
 	in28 := client.NewCrossInput(in1, in2, in27)
 	t.Run("S3MixedCross", func(t *testing.T) {
-		di, err := NewIterator(c, taskDoer, in28)
+		di, err := datum.NewIterator(c, taskDoer, in28)
 		require.NoError(t, err)
 		validateDI(t, di,
 			"/foo11/foo12/", "/foo11/foo2/", "/foo11/foo22/", "/foo11/foo32/", "/foo11/foo42/",
@@ -282,7 +284,7 @@ func TestIterators(t *testing.T) {
 	// in29 is a cross consisting of exclusively S3 inputs
 	in29 := client.NewCrossInput(in27, in27, in27)
 	t.Run("S3OnlyCrossUnionJoin", func(t *testing.T) {
-		di, err := NewIterator(c, taskDoer, in29)
+		di, err := datum.NewIterator(c, taskDoer, in29)
 		require.NoError(t, err)
 		validateDI(t, di, "///")
 	})
@@ -297,9 +299,9 @@ func createTaskDoer(t *testing.T, env *testpachd.RealEnv) task.Doer {
 	go func() {
 		err := taskSource.Iterate(ctx, func(ctx context.Context, input *types.Any) (*types.Any, error) {
 			switch {
-			case IsTask(input):
+			case datum.IsTask(input):
 				pachClient := env.PachClient.WithCtx(ctx)
-				return ProcessTask(pachClient, input)
+				return datum.ProcessTask(pachClient, input)
 			default:
 				return nil, errors.Errorf("unrecognized any type (%v) in test iterators worker", input.TypeUrl)
 			}
@@ -348,7 +350,7 @@ func TestJoinTrailingSlash(t *testing.T) {
 	// Test without trailing slashes
 	input[0].Pfs.Glob = "/(*)"
 	input[1].Pfs.Glob = "/(*)"
-	itr, err := NewIterator(c, taskDoer, client.NewJoinInput(input...))
+	itr, err := datum.NewIterator(c, taskDoer, client.NewJoinInput(input...))
 	require.NoError(t, err)
 	validateDI(t, itr,
 		"/foo-0/foo-0",
@@ -365,7 +367,7 @@ func TestJoinTrailingSlash(t *testing.T) {
 	// Test with trailing slashes
 	input[0].Pfs.Glob = "/(*)/"
 	input[1].Pfs.Glob = "/(*)/"
-	itr, err = NewIterator(c, taskDoer, client.NewJoinInput(input...))
+	itr, err = datum.NewIterator(c, taskDoer, client.NewJoinInput(input...))
 	require.NoError(t, err)
 	validateDI(t, itr,
 		"/foo-0/foo-0",
@@ -381,13 +383,13 @@ func TestJoinTrailingSlash(t *testing.T) {
 	)
 }
 
-func validateDI(t testing.TB, di Iterator, datums ...string) {
+func validateDI(t testing.TB, di datum.Iterator, datums ...string) {
 	t.Helper()
 	datumMap := make(map[string]int)
 	for _, datum := range datums {
 		datumMap[computeStableKey(datum)]++
 	}
-	require.NoError(t, di.Iterate(func(meta *Meta) error {
+	require.NoError(t, di.Iterate(func(meta *datum.Meta) error {
 		key := computeKey(meta)
 		if _, ok := datumMap[key]; !ok {
 			return errors.Errorf("unexpected datum: %v", key)
@@ -401,7 +403,7 @@ func validateDI(t testing.TB, di Iterator, datums ...string) {
 	require.Equal(t, 0, len(datumMap))
 }
 
-func computeKey(meta *Meta) string {
+func computeKey(meta *datum.Meta) string {
 	var key string
 	for _, input := range meta.Inputs {
 		key += input.FileInfo.File.Path
